refactor: Reorganize service and namespace structure
- Moved AuthService and PasswordService from Service to Services folder
- Updated namespace references in Program.cs, Controllers, and other related files
- Corrected import statements to reflect new service location
