+++ API/Controllers/AuthController.cs
@@ -1,5 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
-using API.Service;
+using API.Services;
 using API.Contracts.Responses.Auth;
 using API.Contracts.Responses.Common;
 using API.Contracts.Requests.Auth;
